% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{N.test}
\alias{N.test}
\title{Hypothesis Test and Interval Estimation of the Mean of Normal Population with the variance is known}
\usage{
N.test(x, mu, sigma, alpha, alternative = "two.side")
}
\arguments{
\item{x}{data samples}

\item{mu}{the mean to test}

\item{sigma}{the variance}

\item{alpha}{the confidence level}

\item{alternative}{alternative hypothesis}
}
\value{
test results and confidence intervals
}
\description{
Hypothesis test and interval estimation of the mean when the variance of the normal population is known using R
}
\examples{
\dontrun{
N.test(x=X,mu=0,sigma=1,alpha=0.05,alternative='two.side')
}
}
