% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMetropolis}
\alias{rwMetropolis}
\title{A random walk Metropolis sampler using Rcpp}
\usage{
rwMetropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{the variance}

\item{x0}{the initial value}

\item{N}{the length of the chain}
}
\value{
x  a chain
}
\description{
A random walk Metropolis sampler for generating the standard Laplace distribution using Rcpp
}
\examples{
\dontrun{
rw -> rwMetropolis(2,25,2000)
}
}
