\name{StatComp20067-package}
\alias{StatComp20067-package}
\alias{StatComp20067}
\docType{package}
\title{
  This package provides several functions used in hypothesis test and confidence interval for the mean of a normal population.
}
\description{
  A function for hypothesis test and interval estimation of the mean when the variance of the normal population is known.A function for hypothesis test and interval estimation of the mean when the variance of the normal population is unknown.
}

\author{
Tao Xu

Maintainer: Tao Xu <xutao1996@mail.ustc.edu.cn>
}
\keyword{ package }

