% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{T.test}
\alias{T.test}
\title{Hypothesis Test and Interval Estimation of the Mean of Normal Population with the variance is unknown}
\usage{
T.test(x, mu, alpha, alternative = "two.side")
}
\arguments{
\item{x}{data samples}

\item{mu}{the mean to test}

\item{alpha}{the confidence level}

\item{alternative}{alternative hypothesis}
}
\value{
test results and confidence intervals
}
\description{
Hypothesis test and interval estimation of the mean when the variance of the normal population is unknown using R
}
\examples{
\dontrun{
T.test(x=X,mu=0,alpha=0.05,alternative='two.side')
}
}
